+++ src/entities/questions/questionData.ts
@@ -1,4 +1,5 @@
-import { questionsType, questionType } from '@/entities/questions/questionType'
+import { questionsType } from '@/entities/questions/questionType'
+import { questionType } from '@/entities/questions/questionType'
 
 export const questions: questionsType = [
   {
@@ -28,11 +29,3 @@ export const question: questionType = {
   isHidden: false,
   createdAt: '2024-09-15T10:30:00Z',
 }
-
-// 랜덤 문구 목록
-export const randomDescriptions = [
-  '자유롭게 마음을 표현해보세요! \n지금 시작해볼까요? 😊',
-  '친구들이 뭐라고 답할지 궁금하지 않나요? 🧐 \n지금 바로 퐁! 💌',
-  '퐁~ 하고 던져보세요, \n친구들이 답장해줄 거예요! 😉',
-  '친구들이 당신에게 퐁~ 하고 \n마음을 보낼 준비가 되었답니다! 💕',
-]
